pyproject.toml
[project]
name = "museum"
version = "2.0.0"
description = "Museum of Military Glory named after Oleg Pozdnyakov - Web Application"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Museum Development Team"}
]
keywords = ["django", "museum", "education", "history"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Education",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "django>=5.1.2,<5.2",
    "pillow>=11.0.0,<12",
    "python-dotenv>=1.0.1,<2",
]

[project.optional-dependencies]
dev = [
    "django-debug-toolbar>=4.4.6,<5",
    "ruff>=0.8.0,<1",
    "pytest>=8.3.0,<9",
    "pytest-django>=4.9.0,<5",
    "pytest-cov>=6.0.0,<7",
    "mypy>=1.13.0,<2",
    "django-stubs>=5.1.0,<6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "django-debug-toolbar>=4.4.6",
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-django>=4.9.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.13.0",
    "django-stubs>=5.1.0",
]

# Ruff configuration
[tool.ruff]
target-version = "py311"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "DJ",     # flake8-django
    "RUF",    # Ruff-specific rules
]

ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "DJ001",  # null=True on string-based field
    "RUF012", # mutable class attributes (Django Meta pattern)
    "RUF001", # ambiguous unicode characters (Russian text is OK)
    "ARG002", # unused method arguments (Django command pattern)
    "B007",   # loop variable not used (intentional in some cases)
]

# Exclude a variety of commonly ignored directories
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"*/migrations/*" = ["E501", "DJ001", "RUF012"]
"config/settings*.py" = ["F405"]
"*/tests/*" = ["ARG001", "ARG002"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["Museum", "config"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

# Pytest configuration
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["Museum/tests"]
addopts = [
    "--reuse-db",
    "--cov=Museum",
    "--cov-report=html",
    "--cov-report=term-missing",
    "-v",
]

# MyPy configuration
[tool.mypy]
python_version = "3.11"
plugins = ["mypy_django_plugin.main"]
strict = false
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
no_implicit_optional = true
warn_return_any = false
show_error_codes = true

[[tool.mypy.overrides]]
module = "Museum.*"
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

# Coverage configuration
[tool.coverage.run]
source = ["Museum"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/admin.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
